feat: it reduces create medicine view complexity
