+++ src/hooks/index.ts
@@ -1,2 +1,3 @@
 export * from './getMedicationsParamsReducer'
+export * from './useCreateMedicine'
 export * from './useQueryParams'
